Include Refined armor in `$li`'s tally
The first iteration of the more powerful `$li` command failed to
account for Refined Envoy Armor pieces. This has been rectified.

Up to 6 of the total Perfected Envoy Armor count are added to the
Refined count, which offsets the subtraction of the 6 free pieces,
resulting in the count of crafted Refined pieces.